+++ src/Core/Domain/Models/AnalyticsModel.cs
@@ -11,7 +11,7 @@ public class ProductSalesModel
     public string ProductName { get; set; }
     public string CategoryName { get; set; }
     public string BrandName { get; set; }
-    public int TotalQuantitySold { get; set; }
+    public decimal TotalQuantitySold { get; set; }
     public decimal TotalRevenue { get; set; }
     public int OrderCount { get; set; }
 }
@@ -43,7 +43,7 @@ public class RevenueSummaryModel
     public string Period { get; set; }
     public decimal TotalRevenue { get; set; }
     public int TotalOrders { get; set; }
-    public int TotalItemsSold { get; set; }
+    public decimal TotalItemsSold { get; set; }
 }
 
 public class ProductSalesDetailModel
@@ -53,7 +53,7 @@ public class ProductSalesDetailModel
     public string CategoryName { get; set; }
     public string BrandName { get; set; }
     public decimal CurrentPrice { get; set; }
-    public int TotalQuantitySold { get; set; }
+    public decimal TotalQuantitySold { get; set; }
     public decimal TotalRevenue { get; set; }
     public int OrderCount { get; set; }
     public decimal AverageRating { get; set; }
